requirements.txt
# QuestionForge Dependencies

# Core
pyyaml>=6.0
rich>=13.0.0
click>=8.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# NLP for semantic analysis
nltk>=3.8.0

# Export formats
openpyxl>=3.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
